feat: test for pagination bug and more tracing
There is a bug that causes the pagination to stop working if you try to
go to the prev page when you're on the initial page. Instead, the paging
should just ignore the request and allow the user to attempt to go to
the previous page later.

This also updates and refactors a bunch of methods for clarity and for
making the logs nicer.